requirements.txt
auto_mix_prep==0.2.0
Flask==3.1.2
dash==4.0.0
pandas==3.0.1
numpy==2.4.3
plotly==6.6.0
fiona==1.10.1
geopandas==1.1.3
Requests==2.32.5
tzdata
SQLAlchemy==2.0.43
PyMySQL==1.1.1
python-dotenv==1.2.2
